+++ hosts/nixos/Aarch64LUKS/default.nix
@@ -5,6 +5,39 @@
   ...
 }:
 {
+  imports = [
+    (modulesPath + "/installer/scan/not-detected.nix")
+  ];
+
+  boot.initrd.availableKernelModules = [
+    "usb_storage"
+    "usbhid"
+  ];
+  boot.initrd.kernelModules = [ ];
+  boot.kernelModules = [ ];
+  boot.extraModulePackages = [ ];
+
+  fileSystems."/" = {
+    device = "/dev/mapper/luks-a110e6c3-37e2-4baa-8301-97aab321d7dc";
+    fsType = "ext4";
+  };
+
+  fileSystems."/boot" = {
+    device = "/dev/disk/by-uuid/73CE-06DC";
+    fsType = "vfat";
+    options = [
+      "fmask=0077"
+      "dmask=0077"
+    ];
+  };
+
+  swapDevices = [
+    { device = "/dev/mapper/luks-127bfa81-becb-487f-8ada-53e44c5b375a"; }
+  ];
+
+  boot.initrd.luks.devices."luks-a110e6c3-37e2-4baa-8301-97aab321d7dc".device =
+    "/dev/disk/by-uuid/a110e6c3-37e2-4baa-8301-97aab321d7dc";
+
   # Bootloader.
   boot.loader.systemd-boot.enable = true;
   boot.loader.efi.canTouchEfiVariables = true;