+++ hosts/nixos/X86Pc/default.nix
@@ -1,6 +1,7 @@
 { modulesPath, ... }:
 {
   imports = [ "${modulesPath}/installer/cd-dvd/iso-image.nix" ];
+  networking.hostName = "x86pc";
 
   nixpkgs.overlays = [
     (final: prev: {