+++ variants/cc-generic-aarch64-botany-bay.nix
@@ -1,0 +1,8 @@
+{
+  host = "GenericAarch64";
+  desktop = "gnome";
+  profiles = [ "botany-bay" ];
+  platform = "generic";
+  buildType = "image";
+  configName = "CCGenericAarch64BotanyBay";
+}