+++ variants/generic-aarch64-gnome.nix
@@ -1,0 +1,6 @@
+{
+  host = "GenericAarch64";
+  desktop = "gnome";
+  platform = "generic";
+  buildType = "image";
+}