+++ variants/pinebookpro-gnome.nix
@@ -1,0 +1,6 @@
+{
+  host = "PineBookPro";
+  desktop = "gnome";
+  platform = "rockchip";
+  buildType = "image";
+}