+++ variants/pinebookpro-phosh.nix
@@ -1,0 +1,6 @@
+{
+  host = "PineBookPro";
+  desktop = "phosh";
+  platform = "rockchip";
+  buildType = "image";
+}