+++ variants/pinetab2-installer.nix
@@ -1,0 +1,5 @@
+{
+  host = "PineTab2";
+  platform = "rockchip";
+  buildType = "installer";
+}