+++ variants/pinetab2-phosh.nix
@@ -1,0 +1,6 @@
+{
+  host = "PineTab2";
+  desktop = "phosh";
+  platform = "rockchip";
+  buildType = "image";
+}