+++ variants/pinetab2-plasma.nix
@@ -1,0 +1,6 @@
+{
+  host = "PineTab2";
+  desktop = "plasma";
+  platform = "rockchip";
+  buildType = "image";
+}