+++ variants/x86pc.nix
@@ -1,0 +1,5 @@
+{
+  host = "X86Pc";
+  platform = "x86";
+  buildType = "iso";
+}